refactor(media): extract GeneratedFileMetadataNormalizer to a shared service
PR #947 review (@FExB17): the verbatim `normalizeGeneratedFileMetadata()`
private method was copy-pasted in both WebhookController and
WidgetPublicController. The duplication would silently rot the moment
a third channel handler is added.

- Move the normalisation into a new `final readonly`
  `App\Service\Media\GeneratedFileMetadataNormalizer` service with a
  single `normalize(mixed $fileMeta): ?array{path,type}` method.
- Inject it via constructor DI into WebhookController and
  WidgetPublicController; drop the duplicated private methods.
- Move the unit test out of `tests/Unit/Controller/` into
  `tests/Unit/Service/Media/` and test the service directly, replacing
  the previous ReflectionMethod hack. Expand the non-array branch into
  a data provider that exercises string/int/bool/float/object inputs.

No behaviour change. 12 unit tests pass against the new service; full
backend suite (1521/1521) still green.
